1_Elentra_iLAMS_atm_tool_V7/requirements.txt
streamlit
selenium
pandas
openpyxl
xlrd>=2.0.1 
xlwt

pytest

# pyinstaller

# pip install -r requirements.txt

ollama
faiss-cpu
numpy